requirements.txt
requests
yt-dlp
PyMuPDF
Pillow
tqdm
jsonschema
